Fix gcc tools wrapper for 32-bit archs
The gcc-sdk/* wrappers default on 64-bit tools if neither -m32 nor -m64
is passed. As ar doesn't understand either of these there is no way to
tell it to use the 32-bit variant.

This change defaults on 64-bit tools if the host appears to be
64-bit-ish, 32-bit tools otherwise.

Also fixes a few other stylistic things (error messages go to stderr,
use exec at the end, fix typo in error message).

Change-Id: Ie41aa1716baf247c90ce6952afac37927652fb90